feat: make config-builder optional and re-organise
